finnrissa mind must be studied
the finnrissa mind is truly a marvel to humanity, and subsequently must be studied. the finnrissa brain is capable of developing code for a flight model, physically moving control sticks, design and coding of an entire UI, subsequent porting of everything into VR, but coding a toggle function binded to a keyboard key in csharp is too much for the finnrissa brain to handle. the cerebral cortex of the finnrissa brain closes its borders at such a sight. absorbment of such information is not possible for the machinations and workings of the finnrissa brain to grasp. the finnrissa brain has decided that it has had enough, and would like to go home. therefore, a github push has been made.

AM I TWEAKING??? SAY U SWEAR TO GOD THIS DOESNT WORK! I LITERALLY CODED FOUR OF THESE IN THE PROGRAMMING COURSE! I SWEAR IM NOT TWEAKING. it HAS to be Unity, i refuse to accept my inability to code a toggle switch. but perhaps i did already, since the code you can see is mostly from a youtube video with my exact same issue. my raw brain code isn't there because it did the exact same thing as this and has the same problems as this, but whatevvies i guess.